Added multiple features: see detailed notes
- login name subranges
student accounts are made of the given prefix, and a serial number from 1 to the total count of students. Now instead of the user number range defaulting from 1 to the given student count parameter, the operator can specify a subrange of student numbers to be used with the given prefix
eg with action commands of 'activate' and 'reset_apikey', a given prefix of 'demo_user', a first_student and last_student values of 35 and 55 respectively, only accounts with logins that match the pattern 'demo_user' followed by any number from 35 to 55 will be activated and have their API keys reset.
-custom padding
by default, the numeric suffix to the username is zero-padded to the length of the highest student number. In a class of 100, with prefix 'demo_user_', the first account will be 'demo_user_001'. This value is now settable.
-pave option
with the introduction of subranges, and the existing specifiability of a prefix, the pave option allows all those settings in the command line or config file to be temporarily overridden. Given action commands will once again apply to all non-admin accounts. Note that the 'nuke_from_orbit' command still deletes ALL non-admin accounts, regardless of the existence of the 'pave' modifier. 
-detailed roster output
another action command was added that provides a few more columns of output in the list of non-admin account
-email change
minor change: now if the instructor email doesn't contain a username portion, all student email addresses are provided directly at the given domain, with no '+'. 
eg if the instructor_email is "instructor@domain.tld" then student emails will be "instructor+prefixNN@domain.tld"
but if the instructor_email is simply "@domain.tld" then student emails will be "prefixNN@domain.tld"